seed.py seeds fake product data

+++ database.py
@@ -40,7 +40,7 @@ class Product(db.Model):
     product_id = db.Column(db.Integer, autoincrement=True, primary_key=True)
     title = db.Column(db.String(200), nullable=False)
     description = db.Column(db.Text(), nullable=False)
-    price = db.Column(db.Float(), nullable=False)
+    price = db.Column(db.Numeric(), nullable=False)
     featured = db.Column(db.Boolean, nullable=True)
